Reverted back refactoring.
